Fix drag-and-drop issue on Windows
It turns out there's a race-condition bug in Firefox on Windows wherein
bookmarks.move() will corrupt the internal state of the bookmark tree,
so bookmarks will get fun things like indexes < 0, multiple bookmarks
will get the same index, etc.

This manifests itself by drag-and-drop behaving strangely--bookmarks
getting dropped in the wrong location and such.

The fix is to avoid using move() in parallel, and instead do create()
and update() in parallel.  create() will create a set of dummy
bookmarks, then we query to figure out what order they were created in,
and we update() the ones in the correct slots so the final bookmarks are
all ordered correctly.

This seems to work on Windows; we'll have to see how it works/performs
on Mac...