Use the new tabs.hide() API to hide tabs
The advantage over outright closing a tab is the browser will hang onto
the tab's state so it can be quickly restored.  Obviously this doesn't
sync across browsers (that's why we will continue to use bookmarks), but
it does make it nicer for the common case working within a single
browser.