Don't pass properties with value==undefined to the model
Passing a property that's /undefined/ is different from not passing the
property--as far as the model is concerned, the former is explicitly
assigning the value /undefined/ to the property, while the latter is
making no statement about the property whatsoever.

This fixes an issue where sometimes a stashed tab's title would disappear
when closing the associated browser tab.