Fix refocusAway in a little more general way
Query for pinned tabs so we can more accurately identify the front
tab (and thus we don't have to open a new tab if we're focused on a
pinned tab).  This is also more likely to work if we are asked to
refocus away from a pinned tab for some strange reason.