Don't discard stashed tabs
Instead, wait for some time before discarding them.  We vary the
acceptable age of hidden tabs with the number of active tabs open, so we
discard hidden tabs more aggressively if the user has more active tabs.
We are also more likely to discard tabs that are older, so as to keep
browser memory usage reasonable.