Options page, with option for showing the tab or sidebar
Add a new options page (which is visible in the Firefox Extensions
view), with an option for showing the stash list in a tab or in the
sidebar.  This is the last bit of plumbing needed to make the tab view
work correctly and integrate with the rest of the extension.