Don't use media queries to detect tab-view vs. sidebar-view
This will open up corner cases, since there are some slight semantic
differences between the two--like the ability to collapse in
sidebar-view, and the lack of "stash this tab" in tab view.

Most notably, being able to spontaneously switch back and forth could
allow the user to get into a situation where a folder is collapsed, and
they can't un-collapse it.