Re-enable collapsing in tab view
There are some times when I wish I had it (imperfect as it is).  Still
not sure how to deal with the fact that a collapsed folder in the same
row as an expanded folder takes up the same amount of space (that's just
how flexbox works), but being able to see an overview of all my folders
in tab view is still nice.