Fix thrown exception when tearing off a tab into a new window
When the user tears off a tab into a new window while the stash list is
open, the stash-list model will throw an exception because it sees the
tab is being moved to a different window, but hasn't received a
window-creation event for that window yet.

The fix is to create a new object in the model for the new window, and
populate it with some likely default values on the assumption we will
see more events for that window (with the actual values) later.  This
keeps the model (mostly) consistent while avoiding weird situations
where tabs can exist without windows.