Introduce a new StoredObject module and use it for storing options
The StoredObject module returns plain old JavaScript objects which are
bound to a particular key in browser-local storage, and "magically"
updated when browser-local storage changes.

Refactor options to use this new StoredObject module.

I plan to use the module in a few other places as well, for new
features, but options is a nice simple place to start.