Fix typo preventing ALL background tasks from running
`TabStashTree()` should have been `tabStashTree()`.  Since this line of
code was executed before registering event handlers, timeouts, etc.,
none of the background tasks were happening.  See release notes added in
this commit for user-visible impacts.

This bug was introduced in 2.2, in commit 110c4176 when I added all the
options for controlling what to do with tabs when they are stashed.

Fun fact: the snark-to-code ratio in this commit is 634:1.  (And it's
even higher if you count this commit message.)