vue: Ignore child folders and bookmarks which are undefined
There are a couple cases where the model might legitimately insert
`undefined`s into an array of children--these situations usually happen
when we are restoring a set of recently-closed tabs, and the browser
sends us some events out of order (so we see a tab at position 2 before
we see a tab at position 1, for example).

However, the Vue components weren't handling this correctly--they were
assuming if something was in a `children` array, it was valid.  This led
to crashes while Vue was trying to update the DOM, because it would try
to access properties on undefined objects.

Fix this by ignoring any `undefined`s in folders or folder lists.