Cleanup empty unnamed folders ONLY as a direct result of user actions
Occasionally, when Firefox Sync is enabled, bookmarks will spontaneously
move into their grandparent folder.  From what the Firefox devs tell me,
this is likely happening because the following rough sequence of events
is occurring (assume there are two systems *A* and *B*):

1. *A*: Creates an unnamed, empty folder to hold some tabs
2. *A*: Syncs the empty folder to the cloud
3. *B*: Notices the new empty folder
4. *B*: The Tab Stash garbage collector in index.js runs due to the
   onChanged bookmark event (possibly for some unrelated sync change),
   and deletes the empty folder
5. *A*: Creates bookmarks in the empty folder it created in step 1
6. *B*: Syncs the folder deletion to the cloud
7. *A*: Notices the folder it just created is now deleted, and moves
   the now-orphaned bookmarks into the grandparent folder

So the problem here is that Tab Stash's garbage collector on *B*
observes the empty folder before it has been populated by *A*,
because *A* synced its intermediate state to *B*.  So *B* assumes the
folder is *actually* empty and deletes it.  Then Firefox Sync is left to
resolve the conflict as in step 7.

To fix this, we get rid of the unnamed-empty-folder garbage collector in
index.js, and cleanup the empty folder at the same time we remove the
last bookmark from the folder itself.  This means the empty folder is
always deleted from the same browser that is deleting the last bookmark
in the folder, and we only delete such folders as a result of an actual
user operation (rather than a background cleanup task triggered by
arbitrary bookmark events).

This is a fix/workaround for issue #8, which covers this and potentially
other Sync issues.

See also: https://bugzilla.mozilla.org/show_bug.cgi?id=1549648