npm update
Replace all version numbers in package.json with '*', and use
package-lock.json to record the actual version numbers of everything.
This is, shall we say, somewhat unpleasantly reminiscent of static
linking, but it seems to be the only way to get `npm update` to actually
install new versions of things???

Anyway, pulling in a bunch of updates here as part of the regular
release cycle.