Type-safety cleanups
Tag every coercion with a #undef or #cast tag so they're easy to search
for, and explain why the coercion is happening and how we know it's
safe.

Also, remove some unnecessary coercions, or coercions that could be
trivially made safe (e.g. by adding an if(undefined) or similar).