Support Firefox multi-selection of tabs [#6]
As of Firefox 64, Firefox now supports seleccting multiple tabs at once,
and exposes this info to extensions.

Anywhere we stash "all" tabs, check to see if the user has selected (or
"highlighted") multiple tabs, and stash only the highlighted tabs.
Adjust menu items and tooltip names accordingly.

This also required a small refactoring in stash.js--now the only API to
call is stashTabs() or stashTabsInWindow(), which handles
multi-selection.  bookmarkTabs() is now private--the way to stash tabs
without closing them is to pass `{close: false}` to the appropriate
stash*() function.