style: Fix sidebar issue where some parts of the folder weren't draggable
In the sidebar view, the nav box containing the toolbar buttons for a
folder was expanded to fill the entire width of the sidebar.  However,
this box disables drag-and-drop, so you just couldn't drag a folder
depending on where you started the drag (clicking on the title was OK,
but the empty space next to the toolbar buttons wouldn't work).

Fix the styling so the nav box only takes up as much space as it needs
to (and thus blocks DnD only over the toolbar buttons themselves).