When restoring tabs, only close the front tab if it's not loading
If the front tab is a "new tab", we close it when restoring other tabs,
so the user doesn't wind up with spurious new-tab pages left lying
around.  However, Firefox might still report that a particular tab's URL
is the new-tab page even while it's loading some other URL into that
tab.  So we need to check the status to see if the tab is in the middle
of loading something before we decide to close it.

Fixes #25.