model: Make interface definitions more accurate
Distinguish between a ModelLeaf (which is only a child, e.g. a Tab) and
a ModelParent (which is only a parent, e.g. a Window), and clean up the
model APIs/tests to express exactly what is needed.  This helps to
better-enforce common contracts between Bookmark and Tab model objects.