Cache website favicons instead of asking Google for them [#3]
Asking Google for favicons leaks info about what domains users are
visiting to a company with a mixed record on privacy.  This was
acceptable to get Tab Stash done quickly, but isn't acceptable over the
longer term.

Rather than making requests to a third-party service such as Google,
simply cache all the website icons we've seen in browser-local
storage (indexedDB), and rely on the cache for showing icons for any
tabs that aren't currently open.

The cache is designed with a client/server architecture not because it's
strictly necessary now, but because it will help with some planned
future features.

The test suite for the caching client and service requires some fairly
large changes to the test infra (and other tests) because of adding a
mock indexedDB to the test environment.  The mock indexedDB provides no
APIs for force-draining events, so we have to generally assume that
events are run normally thru the event loop.

This is in contrast to how MockEventDispatcher used to work (where an
explicit drain() call was necessary to fire events).
MockEventDispatcher now follows the same strategy to keep everything
consistent, but this had cascading effects everywhere. :/