Remember the collapsed state of groups in the stash list [#2]
On the local machine, keep track of which groups are open and which are
closed.  This is useful for people who work with different subsets of
their stashes on different machines, and also for people who don't want
to see the list of unstashed tabs showing up all the time (e.g. if they
use the sidebar only to interact with their stashed tabs).