stash: Tell Firefox which window new tabs belong to [#28]
In `restoreTabs()`, we were not passing a windowId to `tabs.create()`,
so the tabs were just going into the active window.  If the user quickly
switches from one window to another while a restore is happening, the
poor, hapless newly-active window will get mobbed by a bunch of tabs
that don't belong there.