Remove bookmark de-duplication feature [#31]
There's a race in which de-duplication might delete things that
shouldn't be deleted, if multiple de-dups are in progress.  The detailed
analysis is in the bug, but the tl;dr is that because dedup always
deletes bookmarks in *other* folders, two dedups running at once may
delete each others' bookmarks.

This can happen even if two dedups are running on two different
computers, because Firefox Sync is very aggressive and may push updates
individually, which get applied while dedup is in progress.  In the
absence of some synchronous communication mechanism between browsers
synced to the same account, I don't see a way to make dedup work
reliably. :/