Don't close the stash list when restoring a single tab [#37]
I've been watching my own behavior/habits with Tab Stash over the last
little bit, and I agree with the reporter that the vast majority of the
time, merely clicking a tab shouldn't cause Tab Stash itself to close.

In all cases, we now keep Tab Stash open, and holding Cmd/Ctrl will
merely decide whether the tab is restored in the background or the
foreground.