Don't stash to old, unnamed groups - try 2
Instead of looking at the modified time, look at the creation time of
the group.  This makes more sense I think because there might be
other/unrelated modifications that happened in the top-most group which
would make it eligible, when it shouldn't be.

Also clamp down on the age of the group--5 minutes is probably a good
starting point.