Be more strict about locating the stash root [#41]
For all the reasons outlined in the comments and the bug, we first look
for a stash root directly under "Other Bookmarks", and only if we don't
find something there do we search the entire tree.

However, this introduces ambiguities if there are multiple "Tab Stash"
folders in the tree.  We try to detect and warn users about these
ambiguities with a link to the wiki to understand and resolve them.