Tighten up model's view of favicons
Ensure the moodel never sees a mix of old/new favicon/url.  I ran across
a weird race when loading/navigating tabs in quick succession (faster
than a human could do) that was resulting in a tab getting the favicon
from whichever site was loaded in the tab before the current site.

This tightens up the model to ensure we really, truly only ever pull
favicons from tabs that have finished loading.

As a nice side effect, it helps reduce some of the flickering that would
happen when loading a new tab or navigating between sites (as tab
updates would come in quick succession).