Improve reliability of imports, esp. with containers [#91]
In some cases, other extensions (especially container extensions) may
immediately close a tab which tries to navigate to certain sites, and
then reopen a new tab headed for that same site.  Re-write how we gather
site info to handle this situation (and generally make the code easier
to understand).

Also, try to do a better job of catching/handling errors that might
occur during import, and ensuring we don't leave the UI in a frozen
state where it's not obvious what's wrong.