siteinfo: Improve reliability of fetching icons/sites
Extend the timeout by quite a bit, and implement separate timeouts for
loading a page (long) vs. a loaded page defining a favicon (short).
Also, we now take what partial info we can find.  Overall this will make
the loading process slower but more likely to find useful data (so it
doesn't need to be redone).