Never stash new/empty tabs [#122]
There is no point to saving them; originally I left them in because
they're useful as a kind of hacky separator, but in practice they are
causing more problems than they solve (because stray new tabs can show
up in groups in a variety of situations).

In particular, with default settings when only a single new-tab page is
open, launching Tab Stash by clicking on the toolbar button will cause a
new group with that single new tab to be created.  Instead it should not
create any groups and just open the UI.