Move stash.ts into the model
Yet another huge refactoring to move all the core functionality that's
lived in `stash.ts` so far into various places in the model.  This
should help improve performance and testability (since the model has
everything we need, we don't need to query the browser).

Another notable improvement is that we can now track changes to the
stash root in real time and update the UI accordingly.  This was not
possible before because the procedure for finding the stash root was
(and still is) complex.