Re-implement drag-and-drop in a more flexible way
Drop the dependency on vue-draggable, which isn't nearly flexible enough
for this particular use case and required a lot of effort/hacking around
to get it to work at all.  Instead, we have a more flexible in-house
drag-and-drop list component which exposes raw DnD events in a
convenient way.

This also lays the groundwork for more flexible DnD options in the
future, like multi-select DnD and dropping links, data, etc. from
external sources.