Refactor the main model to work in terms of DnD operations
We now have more fundamental and flexible methods like
`putItemsInWindow()` to represent DnD operations.  These same methods
can be reused for the "existing" basic operations like stashing all tabs
in a window, restoring a folder, etc.

None of this is tested yet, of course--tests and bugfixes to come in the
following commits. :)