model/bookmarks: Refactor to drop EventfulMap and Index
It turns out that eventual consistency when it comes to
moving/repositioning things doesn't really work, since move events are
not idempotent and cannot be applied out of order.

So we're going back to a stronger-consistency model where we have to
maintain fewer invariants in the bookmark model itself.  And while I'm
at it, tighten up the types used by the bookmark model.