Fix folder-based searching [#118]
I did this as part of the filtering refactoring (to work with
multi-selection), but it was done incorrectly--searching for e.g. "Tab"
would find "Tab" in the name of the stash root and show all the child
folders.

Now we search folder names, but show only the folders themselves, and
leave it to the UI to show/hide filtered items (coming shortly).