Do a better job of tracking the target window
The "target window" is the window whose tabs we should be showing in the
UI, and the place where tabs should be restored to.  For a model in the
background page, the target window should always be the
currently-focused window, but for the UI, it should be the window where
the UI was initially opened.