Don't restore from recently-closed tabs by default [#188]
There is a bug in Firefox which will restore the wrong tab--e.g. we ask
for the tab for example.com and we get example.org instead.

Demote restoring from recently-closed tabs to be an experimental
feature, so we no longer rely on the browser's recently-closed tabs at
all, unless the feature flag is turned on.  If Firefox fixes the bug
later, then perhaps this can be revisited (and enabled by default
once again).