model: Factor the choice of which tabs in a window to stash to its own fn
This splits the mutator and accessor parts of stashTabsInWindow() into
separate functions.  I'm trying to get rid of the various stashTabs*()
functions since they're very similar to some of the other functions
already present in the model (and their weird/optional APIs have proven
to be a source of bugs).