Don't reuse unnamed stashes when saving multiple items [#179, #190]
Refactor and consolidate the various stash*() functions (actually, we
pretty much remove them entirely), and in the process, address the
aforementioned bugs.

Previously, we were appending to a recent unnamed stash even in some
situations where we shouldn't be, such as importing tabs, and when
saving multiple tabs at once.  We only want to append to a previous
stash in situations where we are saving single tabs and haven't
specified a particular destination.

This refactor/bugfix also allows us to considerably simplify model code
and tests, at the expense of a trivial increase in complexity in
UI-facing code.  But it's worth it since we lose ~200 lines of code and
tests without losing any functionality. :D