Don't add duplicate bookmarks to the same folder [#46]
When adding bookmarks/stashed tabs to a folder, don't add them if
there's already one which has the same URL (and title, if the title is
specified by the incoming item).  Instead, move the old one into the new
position.

This is a lighter-weight/less-aggressive version of what Tab Stash used
to do a long long time ago, implemented in a way that should not cause
sync problems.

We intentionally don't de-duplicate across folders, because that might
be surprising--there are cases where users might intentionally want
duplicates in different folders.