Exponential backoff for expensive background operations [#201]
For expensive background operations which might be repeated in a short
period of time, implement exponential backoff so we don't retry them
repeatedly and churn the CPU unnecessarily.

This is particularly relevant for model `reload()` functions which try
to recover from any latent Tab Stash bugs and/or inconsistent results
reported by the browser's APIs.  (Session data such as open/windows tabs
seem to be a particularly bad culprit here.)

If the user gets into such a situation, there's not much we can do to
fix it, but we can at least use less CPU while working around it.