Don't store deleted items in memory unless absolutely necessary [#172]
For a variety of reasons, the background page has an instance of the
deleted-items model.  Until now, the deleted-items model would, upon
receiving an event that a new deleted item had been added, immediately
store that item in an in-memory array of deleted items. (Importantly,
the item was always added at the *beginning* of the array, shifting all
subsequent items right by one.)

The problem is that if the user was not restarting their browser and/or
had a lot of deleted items, these items would accumulate in memory.
Each additional item would require shifting the *entire* array, so in
aggregate, saving deleted items turned out to be `O(N^2)`.

When it came time to garbage-collect old deleted items (once a day or on
browser restart), old items would be *removed* from this array, again
shifting things around to the tune of `O(N^2)`.  Why `O(N^2)` if we're
deleting the *oldest* items, you ask?  Well, `onDelete()` doesn't assume
the entry is at the old end of the array, and *filters the entire array*
for each set of deletions.  (We delete in batches of 50, so this isn't
quite as bad as it sounds, but it's still bad.)

So the fix is:

Don't store deleted items in the model unless the UI has actually asked
for some entries to show.  Now, when we get an event saying a deleted
item has been added, we just tell the model it's not fully-loaded
anymore (unless the item is newer than the oldest item already present
in the model--if any are present at all).  This keeps the background
page's model state nice and empty so it doesn't matter if shuffling the
array around is `O(N)`, `O(N^2)` or `O(N!)`).

With this change, we don't *need* to make `onDelete()` faster.  Before,
`onDelete()` should have been fast to deal with GC-triggered deletions.
Now, deletions relevant to the model are almost *always* in situations
where the user is deleting an arbitrary item somewhere in the middle of
the array, and we can assume the user will have only a reasonable number
of deleted items loaded in the UI (and thus the model).