KVSClient: Notify users and reconnect when disconnected from the service
In certain situations, client NanoPorts can be spontaneously
disconnected from their counterparts in the background page.  If this
happens to a KVS, we need to reconnect and notify users of the KVS that
events may have been lost.  Users can then take appropriate actions to
re-sync themselves with the actual state of the KVS as needed.