kvs/client: Retry operations when a comm issue occurs [#223]
If a KVSClient gets disconnected while a request is in flight, that
request will be dropped by the now-defunct NanoPort.  Similarly, if a
request gets dropped (and times out), the client has no idea if it was
actually completed or not.  In both cases, we can retry the request on
the assumption it's a transient issue.