Rate-limit tab loading to avoid slowing down the user's system [#129]
If Tab Stash creates a tab, create it in a `discarded` state, and then
try to load it later in the background.  But, as part of loading it
after the fact, make sure we never try to load more than about 4 at a
time, to avoid slowing down the user's system.  Otherwise, restoring
very large stashes may cause everything to grind to a halt.