Refactor editing of folder names to allow editors to be used elsewhere
Right now, editable-label assumes that it will be activated/deactivated
by clicking on the text itself, but it would be nice to be able to use
it in other situations where editing is triggered by some other
mechanism.  Rewrite editable-label as an async-text-input component
which can be used in this way.

While I'm here, fix an issue that was causing folder renaming to flicker
briefly back to the old name (due to improper use of async/promises).