Fix a number of bugs related to favicon handling and title storage
We would sometimes improperly invalidate entries in the persistent
favicon cache just because we had stale data in the local cache.  Now we
make more of an attempt not to do this.

We also try not to update the favicon cache unless something has
actually changed.  This should improve performance.