Fix issue dragging consecutive items [#295]
The DnD list was losing track of which items were at which positions in
the list, because of some oddities in how Vue.js maintains references to
DOM elements.  Rather than relying on Vue to do this for us, we can
figure out which DOM element we need by looking at the incoming DOM
events themselves.