dnd-list: Improved method of debouncing
The old method would sometimes cause the ghost to get "stuck" in the
wrong place; the new method relies on the actual visual dimensions of
the list and so it's a little more intuitive to the user.

The new method is simply to resize the ghost to the same dimensions as
the item it's about to replace; that way we don't wind up with it
vibrating back and forth between two positions (e.g. if moving a small
ghost into a large item causes the large item to shift in such a way
that the ghost is no longer under the cursor in its new location).