perf: Incrementally compute the number of selected bookmarks
The computed() handler for `model.bookmarks.selectedCount` was the
single largest consumer of CPU during a tab import(!).  This seems to be
because it was walking the entire stash on every hierarchy change--e.g.
adding to or removing tabs from the stash.  This made adding tabs to the
stash be O(N^2).

This is something that can easily be incrementally updated in constant
time, bringing tab addition back down to O(N).  So instead of using
computed() for `selectedCount`, make it be a ref() and manually update
the count at relevant points.