release builds: Remove sourcemaps and don't minify
Counter-intuitively, it turns out that this both drastically reduces the
size of the built artifacts AND is also more reliable in getting useful
stack traces.  Source maps don't always seem to work for reasons I've
never been able to figure out, and minifying the build means that a lot
of things in the stack trace are single-letter names instead of their
actual names.