dnd-list: Provide both displacing and non-displacing ghost styles
The "displacing" ghost style is what Tab Stash has traditionally
done--when the drag-and-drop ghost appears or moves, it "displaces"
items in the list.  So it visually acts like a placeholder that moves
items around it to show where the dragged item will be dropped.

The new "non-displacing" ghost style doesn't cause items to shift when
it appears/disappears.  Instead, it just draws a line between items to
indicate where the dragged item will be dropped.

We need both styles to eventually support nested folders--the displacing
ghost won't work with nested folders, because you can't drag into a
nested folder--the folder just moves out of the way.  The non-displacing
ghost, on the other hand, won't work with the top-level stash list,
because it's a tiled view--adding the ghost as an overlay will
completely mess up the flexbox.