kvs: Decrease batch size of KVSCache [#335]
On slower machines or machines that are doing many things, even 100
entries at a time might be too much, especially if they're large entries
(like favicons).  Fetching 25 at a time doesn't noticeably slow things
down on my system, but it will cut the request size down to a quarter of
what it was, which might help here.