model/tabs: In events, downgrade unknown-tab errors to warnings [#338]
Firefox seems to enjoy telling us about things that happen to tabs that
it also pretends don't exist--for example, devtools tabs.  When we get
an event for a tab that we don't recognize, ignore it and log a warning
instead of crashing--it's more likely to be a Firefox bug than a Tab
Stash bug.