Fix UI crash when stashing from the browser toolbar
If the UI opens mid-stash, we will sometimes see bookmark events for
children inside folders that we don't yet know about, and we will crash
because the bookmark model doesn't know where to put the child.  Avoid
this by ignoring events while loading the bookmark model, on the
assumption that the model will eventually converge (once events settle
down).