Fix performance regression when updating the UI
Somehow the `selectedCount` and other stats were triggering re-renders
of the entire page, probably because each folder component was checking
whether to show "selected" or "non-selected" toolbar buttons, and the
`selectedCount` was recalculated every time a window or tab changed.

We can do this recalculation by not triggering downstream updates, which
is what the new `computedLazyEq()` ref type does.  (Incidentally, Vue
has something similar internally, but it is not exposed for public
consumption, sadly.)