Fix selection of open tabs
The `selectedCount` computed property wasn't firing, and therefore the
rest of the UI didn't notice that tabs were being selected.  This
happened because the targetWindow was set before the window was actually
loaded in the model, but fetching the window from the model was not,
itself, reactive.  We hack around this by making targetWindow always
refer to the window itself, and assigning window objects to things
targetWindow depends on when those windows are actually populated.