Use global variables instead of Vue injections
Honestly, they're easier to manage since they're all in one place and
there aren't that many of them.  We also don't need any of the more
advanced features of injections (like being able to change the value of
a thing partway down the hierarchy), so the performance overhead of
injection just isn't worth it.

This also makes it more clear/readable as to where to find state (and
there will be more UI state, separate from the model, soon).