On restore, Delete bookmarks before closing the active tab [#377]
When restoring and removing tabs from the stash, during the "restore"
phase, we close the active tab if it's a "blank" tab (i.e. the new-tab
page, the homepage,`about:blank`, and so on).  Only after closing the
blank tab do we then delete the stashed tabs from the bookmark store.

The problem is that Tab Stash itself might be the "blank" tab (e.g. if
Tab Stash is set as the user's homepage).  If Tab Stash closes itself,
it cannot finish cleaning up the bookmarks for the restored tabs.

This patch deletes the bookmarks (if needed) before attempting to close
the active "blank" tab, so the bookmarks will always get cleaned up
regardless of whether Tab Stash is considered to be a "blank" tab.