Allow stashing of redirected URLs
If the user unstashes and then re-stashes a privileged URL, the internal
Tab Stash redirection page might get saved instead of the actual URL the
user wants.  Check for and explicitly allow (re-)stashing of redirection
pages, and save the original URL instead.

This has the added benefit that we don't put Tab Stash-specific
droppings in the user's bookmarks, either.