Don't show new-version notification on fresh installs
The new-version banner was briefly showing on a fresh install, because
`undefined` is less than whatever the current version is, and the
background page hasn't had a chance to actually write the version number
to local storage yet.

If we see `undefined` for the last_notified_version, just assume we
don't need to notify the user of a new version.