Track hidden tabs using session data
Whenever we hide a tab, tag it with a marker in the browser session data
so we can tell it was Tab Stash that hid it.  This allows us to more
accurately identify tabs that need to be cleaned up later.

We were previously cleaning up hidden tabs by looking only at what URLs
were removed from the stash and closing the corresponding tabs.
However, that doesn't always work, because the tab's URL could
spontaneously change (e.g. a user getting logged out), and then we
wouldn't think it belonged to us anymore.

Now, if we see a tab with the Tab Stash marker but with a URL that isn't
stashed, we know it's one of ours and we can close it, regardless of why
the URL is unrecognized.

Closes #425.