Nicer drag-and-drop behavior
Pay attention to where inside the hovered item the mouse cursor is, to
decide if the dragged item should be dropped above or below the hovered
item. This leads to a nicer/more intuitive feel for dropping items,
especially for appending to a list.

This improves, but does not fix, #524.