Fix crash when tab is closed and then opened quickly
As the code comment says.  I tested this manually on my system--was able
to repro the crash without the fix, and verify the fix avoids the crash.

This is technically a workaround for an upstream Firefox bug, since
Firefox sends us a tab-creation event with an index that assumes the
closed tab is still open.

Closes #537.