Fix selection of new active tab when stashing the active tab
When we stash the active tab, we need to choose a new tab to become the
active tab.  We were doing this calculation improperly, because we were
slicing into a filtered list of tabs rather than the actual list of tabs
present in the window.  This would cause us to choose a tab that was
offset by the number of pinned+hidden tabs in the window.

Fix this and add some tests to make sure we choose the correct tab.

Closes #369.