Turn drag-and-drop behavior into self-contained functions
Now DnD behavior can be applied to any arbitrary HTML element in a way
that makes only minimal use of Vue's reactivity and avoids the overhead
of an entire component.  This is useful for adding DnD behavior in a way
that doesn't impact the DOM.

We also allow for more flexibility in what drop targets accept--they can
decide, at validation time, where to accept a particular drop, and they
can look at more than just the available data types to decide whether to
accept or not.