Respect search terms when checking for the default stash destination
When stashing individual tabs, the logic to check if we already have a
"default" stash destination was not taking into account the behavior
introduced in a1ffc9b4 to use the search term as the folder's title.

We now consider the topmost folder a "default destination" if its title
matches the currently active search term.  This should prevent the
repeated creation of new folders as described in the bug.

Closes #581.