import: Don't fail even if we can't fetch some URLs
Just fetch as many URLs we can, collect the list of URLs that we can't,
and report those to the user at the end.  But leave the imported tabs
alone, so users can fix things up manually.

Most of the time, only a few URLs in the big import will fail, and it's
still useful to have the import at large succeed, even if we don't have
titles/icons for a few things.

Closes #500.