popup: Pass the current selection thru when opening the UI in a tab
When the user is in the popup UI and has some items selected, we should
still let them open the UI in a tab. Moreover, when opening the tab, we
should carry over their selection from the popup, so they can continue
doing whatever they were in the middle of.