folder menu: Hide the list of tabs to stash by default
Make this "Stash to ..." list of currently-open tabs in the folder menu
hidden behind a `<details>`. This makes the menu more compact, making
options below the list more accessible, and avoids an issue where the
menu doesn't have enough room to display even one tab in the list (thus
making the list unusable).