Make search-box focusing more reliable in popup view
Make the focus behavior more reliable for the popup view by adding a
short delay, to avoid Firefox-internal things which mess with the focus.

Also, try to focus the search box automatically when opening the
sidebar. This doesn't actually work because Firefox completely ignores
attempts to move the focus to the sidebar when it opens. But if
Firefox's behavior ever changes, Tab Stash will behave appropriately.

Closes #622.